Work around the issue of SIGKILL
The containerd CRI plugin sends SIGKILL to a single process of
a container when timeout is not set. If the process has multiple
child processes and only the parent process is killed, the child
processes are alive, and stdio/stderr streams are not closed.
Due to this behavior, "crictl stop" never finishes.
To work around this issue, the all flag is turned on for SIGKILL.

https://github.com/containerd/containerd/blob/v1.5.8/pkg/cri/server/container_stop.go#L167
Signed-off-by: Yohei Ueda <yohei@jp.ibm.com>